turns out johny is bad
